test(loading): expose test helpers
