+++ packages/ubuntu_desktop_installer/test/installer_test.dart
@@ -10,7 +10,7 @@ import 'package:ubuntu_desktop_installer/services.dart';
 import 'package:ubuntu_wizard/utils.dart';
 
 import 'install/install_model_test.mocks.dart';
-import 'loading/loading_model_test.mocks.dart';
+import 'loading/test_loading.dart';
 import 'locale/test_locale.dart';
 import 'theme/theme_page_test.mocks.dart';
