+++ packages/ubuntu_desktop_installer/test/loading/loading_page_test.dart
@@ -1,34 +1,15 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
-import 'package:ubuntu_desktop_installer/pages/loading/loading_model.dart';
 import 'package:ubuntu_desktop_installer/pages/loading/loading_page.dart';
 import 'package:yaru_test/yaru_test.dart';
 
 import '../test_utils.dart';
-import 'loading_page_test.mocks.dart';
+import 'test_loading.dart';
 
-@GenerateMocks([LoadingModel])
 void main() {
-  LoadingModel buildModel({Duration? delay}) {
-    final model = MockLoadingModel();
-    when(model.init())
-        .thenAnswer((_) => Future.delayed(delay ?? Duration.zero));
-    return model;
-  }
-
-  Widget buildPage(LoadingModel model) {
-    return ProviderScope(
-      overrides: [loadingModelProvider.overrideWith((_) => model)],
-      child: const LoadingPage(),
-    );
-  }
-
   testWidgets('init & advance', (tester) async {
-    final model = buildModel(delay: const Duration(seconds: 3));
-    await tester.pumpWidget(tester.buildApp((_) => buildPage(model)));
+    final model = buildLoadingModel(delay: const Duration(seconds: 3));
+    await tester.pumpWidget(tester.buildApp((_) => buildLoadingPage(model)));
 
     verify(model.init()).called(1);
     expect(find.byType(LoadingPage), findsOneWidget);
@@ -39,7 +20,7 @@ void main() {
 
   testWidgets('disabled buttons', (tester) async {
     final model = MockLoadingModel();
-    await tester.pumpWidget(tester.buildApp((_) => buildPage(model)));
+    await tester.pumpWidget(tester.buildApp((_) => buildLoadingPage(model)));
 
     expect(find.button(tester.ulang.previousLabel), isDisabled);
     expect(find.button(tester.ulang.nextLabel), isDisabled);
